fix unsigned int

+++ smacc2_client_library/nav2z_client/custom_planners/backward_global_planner/src/backward_global_planner/backward_global_planner.cpp
@@ -267,7 +267,7 @@ nav_msgs::msg::Path BackwardGlobalPlanner::createPlan(
   auto costmap2d = this->costmap_ros_->getCostmap();
   for (auto & p : plan)
   {
-    unsigned int mx, my;
+    uint64_t mx, my;
     costmap2d->worldToMap(p.pose.position.x, p.pose.position.y, mx, my);
     auto cost = costmap2d->getCost(mx, my);
